Transceiver52M: Re-enable autoconf x86 arch detection
commit c13f8bde8f1d00f44cc854fa77b4196a733d7808 "merged in commercial
openbts" removed AC_CONFIG_MACRO_DIR in configure.ac, which in effect
disabled SSE support on environments where the detection macro, AX_EXT,
was not already present.

Rather than rely on AX_EXT be present in the distribution, use the
version that is already included with OpenBTS by re-enabling the config
m4 directory.

Signed-off-by: Thomas Tsou <tom@tsou.cc>